Change preview deplay on cell views and annd file hierarchy
