+++ PlayerMusicLyricAnimationPOC/Model/Music.swift
@@ -17,25 +17,3 @@ struct Music: Identifiable, Hashable {
 	let lyrics: [String]
 }
 
-// MARK: Music data
-
-extension Music {
-	static let musicSample: Music = Music(
-		artist: "Eminem",
-		cover: "eminemCover",
-		title: "The Real Slim Shady",
-		lyrics:
-			[
-				"May I have your attention, please ?",
-				"May I have your attention, please ?",
-				"Will the real Slim Shady please stand up ?",
-				"I repeat, will the real Slim Shady please stand up ?",
-				"We're gonna have a problem here...",
-				"May I have your attention, please ?",
-				"May I have your attention, please ?",
-				"Will the real Slim Shady please stand up ?",
-				"I repeat, will the real Slim Shady please stand up ?",
-				"We're gonna have a problem here..."
-			]
-	)
-}
